contactsapi/requirements.txt
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
django-bootstrap4==2.3.1
django-cors-headers==3.10.0
djangorestframework==3.12.2
drf-yasg==1.20.0
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imutils==0.5.3
inflection==0.5.1
jsonschema==3.2.0
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
Pillow==8.0.1
Pygments==2.7.3
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2020.5
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.55.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
gunicorn==20.1.0

